pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.14.1",
    "langchain>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "langgraph-supervisor>=0.0.31",
    "mcp>=1.25.0",
    "motor>=3.7.1",
    "oracledb>=3.4.1",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "watchdog>=6.0.0",
]
